Metrics changes + tests
